requirements.txt
Django==1.10.6
django-crispy-forms==1.6.1
django-mptt==0.8.7
django-pagedown==0.1.3
django-registration-redux==1.7
django-avatar==4.0.0
djangorestframework==3.6.3
Pillow==4.1.1
uWSGI==2.0.14
